requirements.txt
Django==5.1.1
django-allauth==65.8.1
psycopg2-binary==2.9.10